add contacts.tsx and footer-paragraphs.tsx + refact footer.tsx
